// prev/packages/tree/src/types.ts
export type TreeItemId = string;
export type ExpandedIds = readonly TreeItemId[];
export type SelectedIds = readonly TreeItemId[];

export interface TreeItemIds {
  /**
   * The unique identifier for an item within the tree. This is used to be able
   * to link tree items together with parent items as well as selected/expanded
   * logic.
   */
  itemId: TreeItemId;

  /**
   * This id is used to be able to link tree items together with children and
   * parents. If the `parentId` is set to `null`, the tree item will appear at
   * the root of the tree. Otherwise, the tree item will appear as a child of
   * the referenced tree item by `itemId`
   */
  parentId: null | TreeItemId;
}

export interface BaseTreeItem extends TreeItemIds, ListItemChildrenProps {
  /**
   * Boolean if the provided item is a custom element and will not be rendered
   * in a `TreeItem` component. This is useful if you want to be able to render
   * `Divider` or `ListSubheader` components within a tree since they _should_
   * be able to be rendered without any of the tree functionality.
   */
  isCustom?: boolean;
}

/**
 * This is just a type I use internally for handling tree items with unknown
 * keys.
 *
 * @internal
 */
export interface UnknownTreeItem extends BaseTreeItem {
  [key: string]: unknown;
}

/**
 * This type represents the data that can be handled by the `Tree` component.
 * This is really just a Map of `itemId -> item` for quick lookups for the logic
 * within the tree.
 */
export type TreeData<T extends BaseTreeItem> = Record<TreeItemId, T>;